requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.7
aiohttp>=3.9
PyYAML>=6.0
SQLAlchemy>=1.4,<2.0
alembic>=1.12
pydantic>=2.0
pytest>=7.4
pytest-asyncio>=0.21
async-timeout>=4.0